优化代码，使更简洁

+++ vsconvert.py
@@ -26,8 +26,10 @@ class VsConvert(object):
             self.groups = {}
         if filepath is None:
             self.action_cfg = ReadCfg().readcfg(action_file)
+            self.cfgfilepath = action_file
         else:
             self.action_cfg = ReadCfg().readcfg(filepath)
+            self.cfgfilepath = filepath
 
     def __check_acton_hostcfg(self, hostobj):
         hostnames = []
@@ -42,37 +44,42 @@ class VsConvert(object):
             return True, []
 
     def __action_func_inner(self, hostname, modelname, pub_param, param):
-        self.mylog.info("------模块:{mod},主机:{host}" .format(host=hostname, mod=modelname))
-        if not os.path.isdir(pub_param["source"]):
-            self.mylog.cri("public source目录不存在")
-            raise Exception("public source目录不存在")
+        self.mylog.info("------模块:{mod},主机:{host}".format(host=hostname, mod=modelname))
 
         m = importlib.import_module("model." + modelname + "." + modelname)
         m.ModelClass(self.mylog).action(hostname, pub_param, param)
 
-    def action_func(self):
-        # 检查hosts 配置是否有错误的，如果有错误，则不运行
+    def __prerun_check(self):
+        # action文件，public中的source配置的目录不存在则退出
+        pub_param = self.action_cfg["PUBLIC"]
+        if not os.path.isdir(pub_param["source"]):
+            self.mylog.cri("public source目录不存在")
+            raise Exception("public source目录不存在")
+
+        # action文件，public中的dest配置的目录已存在且选择删除则删除
+        if os.path.isdir(pub_param["dest"]) and input(f"目录 {pub_param['dest']} 已存在，是否删除(y/n) ") == "y":
+            shutil.rmtree(pub_param["dest"])
+        os.makedirs(pub_param["dest"])
+
+        # 检查hosts 配置是否有错误，如果有错误，则不运行
         hostobj = hosts(self.mylog, self.groups, self.action_cfg["PUBLIC"]["source"])
         checkrz, rzlist = self.__check_acton_hostcfg(hostobj)
         if not checkrz:
             self.mylog.cri(f"action 配置文件中hosts错误：{rzlist}")
             raise Exception("action 配置文件hosts配置错误")
 
-        # 开始-测试用，正式发布时去掉
-        pub_param = self.action_cfg["PUBLIC"]
-        if os.path.isdir(pub_param["dest"]):
-            t = pub_param["dest"]
-            if input(f"目录 {t} 已存在，是否删除(y/n) ") == 'y':
-                shutil.rmtree(pub_param["dest"])
-        os.makedirs(pub_param["dest"])
-        # 结束-测试用，正式发布时去掉
+        return hostobj
 
+    def action_func(self):
+        hostobj = self.__prerun_check()
+
+        self.mylog.green(f'############开始任务{self.cfgfilepath}############')
         for action in self.action_cfg["ACTION"]:
             # 遍历hosts列表
             hostname_list, err_host_list = hostobj.get_host_name(action["hosts"])
             for task in action["tasks"]:
                 # 遍历task列表
-                self.mylog.info('#########执行任务：{task}#########'.format(task=task["name"]))
+                self.mylog.info('*********执行任务：{task}*********'.format(task=task["name"]))
                 # 模块信息
                 for modelname, param in task.items():
                     if modelname == "name":
@@ -82,4 +89,4 @@ class VsConvert(object):
                         continue
                     for hostname in hostname_list:
                         self.__action_func_inner(hostname, modelname, self.action_cfg["PUBLIC"], param)
-        self.mylog.info('#########所有任务完成#########')
+        self.mylog.green('#########所有任务完成#########')
